marketing/coupang_partners/requirements.txt
flask>=3.0.0
firebase-admin>=6.2.0
google-cloud-firestore>=2.13.0
google-api-python-client>=2.100.0
anthropic>=0.96.0
python-dotenv>=1.0.0
requests>=2.31.0
